Add time ci method
npm run time and yarn time are available, yarn time is recalculating performance test-like for union (see unionTime), TIMES.md is added, yarn time shows table with result and rewrite TIMES.md, comparison is carried out for 1000 elements, the mean is calculating 1000 times to reduce side effects of laptop and computer. The performance test for union is tested for 3 cases: 0, 50 and 100% of elements are the same.

The data is generated in function unionTime, other methods should be placed same way. Maybe use *.json for this reason.
node -v: 8.9.4
npm -v: 5.6.0